+++ src/server/js/index.js
@@ -7,3 +7,4 @@ import '../../common/js/index.js';
 import './check_npms.js';
 import './functions.js';
 import './methods.js';
+import './publish.js';