+++ ChangeLog.md
@@ -12,6 +12,7 @@
     - Reviewed all functions to use async versions and return Promises (todo #3)
     - New AccountsTools.byEmail() function
     - New AccountsTools.byUsername() function
+    - New AccountsTools.emails publication
 
 ### 2.0.0